+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Security
 -->
 ## Unreleased
+### Fixed
+- [RT#186373]: in some situations, an article can be accepted several times. Now, the first date of acceptance will be taken into account.
 ### Changed
 - Bump from TinyMCE 4 to TinyMCE 6
 - [278](https://github.com/CCSDForge/episciences/issues/278): In case an article has been refused: new option to allow forward automatically the message sent to the authors explaining the final decision made by the editor in charge.